+++ R/usethis-package.R
@@ -4,7 +4,8 @@
 #' @importFrom rlang := %|% %||%
 "_PACKAGE"
 
+# The following block is used by usethis to automatically manage
+# roxygen namespace tags. Modify with care!
 ## usethis namespace: start
 ## usethis namespace: end
 NULL
-