+++ R/git-utils.R
@@ -260,6 +260,11 @@ check_branch <- function(branch) {
 
 check_branch_pulled <- function(branch = git_branch_name(), use = "git pull") {
   remote <- git_branch_tracking_FIXME(branch)
+  if (is.null(remote)) {
+    ui_done("Local branch {ui_value(branch)} is not tracking a remote branch.")
+    return(invisible())
+  }
+
   ui_done(
     "Checking that local branch {ui_value(branch)} has the changes \\
      in {ui_value(remote)}"