+++ NEWS.md
@@ -1,5 +1,7 @@
 # usethis (development version)
 
+* `create_from_github()` sets remote tracking branch of `master` to `upstream/master`, when it creates (and clones) a fork (#792).
+
 * `use_pipe()` gains a logical `export` argument, so it can do the setup necessary to use the pipe operator when it is re-exported (`export = TRUE`, which is the default and preserves the previous behaviour) and when it is not (`export = FALSE`) (#783).
 
 * `use_circleci()` creates a `.circleci/config.yaml` config file for CircleCI